refactor: move interface to resolver
